Django-nextjs-backendApi/requirements.txt
Django
django-ninja
django-cors-headers
django-ninja-jwt[crypto]
pydantic[email]
rav